add size_t to problems variant return type. refactor all the problems returning non-negative result

+++ problems/inc/AoC/2015/problem_01.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstddef>
 #include <iosfwd>
 
 namespace AoC_2015
@@ -8,7 +9,7 @@ namespace AoC_2015
 namespace problem_01
 {
 int solve_1( std::istream& input );
-int solve_2( std::istream& input );
+size_t solve_2( std::istream& input );
 }  // namespace problem_01
 
 }  // namespace AoC_2015
