+++ problems/inc/AoC/2015/problem_25.h
@@ -1,15 +1,15 @@
 #pragma once
 
+#include <cstddef>
 #include <iosfwd>
-#include <string>
 
 namespace AoC_2015
 {
 
 namespace problem_25
 {
-int solve_1( std::istream& input );
-int solve_2( std::istream& input );
+size_t solve_1( std::istream& input );
+size_t solve_2( std::istream& input );
 }  // namespace problem_25
 
 }  // namespace AoC_2015
