+++ problems/inc/AoC/2016/problem_03.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstddef>
 #include <iosfwd>
 
 namespace AoC_2016
@@ -7,8 +8,8 @@ namespace AoC_2016
 
 namespace problem_03
 {
-int solve_1( std::istream& input );
-int solve_2( std::istream& input );
+size_t solve_1( std::istream& input );
+size_t solve_2( std::istream& input );
 }  // namespace problem_03
 
 }  // namespace AoC_2016
